DEU-4
Implemented dummy repo. Fixed header typo. Added decoders
